requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
huggingface_hub>=0.24.0
fastapi>=0.110.0
uvicorn>=0.23.0
pydantic>=2.0.0
